Deleted all Async Tasks to put Executors with Runnables instead, adjusted Settings Activity to System bars and fixed all music ids to long
