--- app/src/main/java/com/maybe/maybe/database/async_tasks/OnFillDbAsyncTaskFinish.java
@@ -1,5 +1,0 @@
-package com.maybe.maybe.database.async_tasks;
-
-public interface OnFillDbAsyncTaskFinish {
-    void onFillDbAsyncTaskFinish();
-}
