--- app/src/main/java/com/maybe/maybe/database/async_tasks/OnSaveCurrentListAsyncTaskFinish.java
@@ -1,9 +1,0 @@
-package com.maybe.maybe.database.async_tasks;
-
-public interface OnSaveCurrentListAsyncTaskFinish {
-    void onSaveCurrentListAsyncTaskStart(int max);
-
-    void onSaveCurrentListAsyncTaskProgress(int progress);
-
-    void onSaveCurrentListAsyncTaskFinish();
-}
