--- app/src/main/java/com/maybe/maybe/database/async_tasks/OnSelectAlbumAsyncTaskFinish.java
@@ -1,7 +1,0 @@
-package com.maybe.maybe.database.async_tasks;
-
-import java.util.List;
-
-public interface OnSelectAlbumAsyncTaskFinish {
-    void onSelectAlbumAsyncFinish(List<Object> objects);
-}
