--- app/src/main/java/com/maybe/maybe/database/async_tasks/playlist/PlaylistAsyncTaskLongResponse.java
@@ -1,7 +1,0 @@
-package com.maybe.maybe.database.async_tasks.playlist;
-
-import java.util.List;
-
-public interface PlaylistAsyncTaskLongResponse {
-    void onPlaylistAsyncTaskLongFinish(List<Long> longs);
-}
