--- app/src/main/java/com/maybe/maybe/database/async_tasks/playlist/PlaylistAsyncTaskNullResponse.java
@@ -1,5 +1,0 @@
-package com.maybe.maybe.database.async_tasks.playlist;
-
-public interface PlaylistAsyncTaskNullResponse {
-    void onPlaylistAsyncTaskNullFinish();
-}
