+++ app/src/main/java/com/maybe/maybe/database/runnables/IFillDbRunnable.java
@@ -1,0 +1,5 @@
+package com.maybe.maybe.database.runnables;
+
+public interface IFillDbRunnable {
+    void onFinish();
+}
