+++ app/src/main/java/com/maybe/maybe/database/runnables/IMusicRunnable.java
@@ -1,0 +1,7 @@
+package com.maybe.maybe.database.runnables;
+
+import java.util.List;
+
+public interface IMusicRunnable {
+    void onFinish(List<Object> objects);
+}
