+++ app/src/main/java/com/maybe/maybe/database/runnables/ISaveCurrentListRunnable.java
@@ -1,0 +1,9 @@
+package com.maybe.maybe.database.runnables;
+
+public interface ISaveCurrentListRunnable {
+    void onSaveCurrentListRunnableStart(int max);
+
+    void onSaveCurrentListRunnableProgress(int progress);
+
+    void onSaveCurrentListRunnableFinish();
+}
