+++ app/src/main/java/com/maybe/maybe/database/runnables/playlist/IPlaylistRunnableNull.java
@@ -1,0 +1,5 @@
+package com.maybe.maybe.database.runnables.playlist;
+
+public interface IPlaylistRunnableNull {
+    void onFinishNull();
+}
