+++ app/src/main/java/com/maybe/maybe/database/runnables/playlist/IPlaylistRunnableObject.java
@@ -1,0 +1,9 @@
+package com.maybe.maybe.database.runnables.playlist;
+
+import com.maybe.maybe.fragments.category.ListItem;
+
+import java.util.ArrayList;
+
+public interface IPlaylistRunnableObject {
+    void onFinish(ArrayList<ListItem> listItems);
+}
